requirements_agent.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai
pydantic>=2.0.0
qdrant-client>=1.7.0
protonx>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-logging>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
streamlit
